dvbvideosink: send VIDEO_CONTINUE only when decoder is paused
Extra VIDEO_CONTINUE ioctl is send on change state
from paused to playing, when decoder is not paused.

This state change is happening after we flush pipeline
from playing state (flush seek).
Flushing creates flush start event, which we pass in
dvbvideosink to basesink. Basesink will lost current state
in flush start, since we have async state changes enabled and
re-commits it again when first buffer is prerolled, creating
paused to playing state change while decoder is not paused.

We need not to send extra VIDEO_CONTINUE when decoder is
not paused, because it cancels playback in rate > 1.0 on
some receivers

Note:
This patch is related directly to previous one, since
before it, we were not passing flush start event from
playing state to basesink